Check whether there is content before initialization
This is a fix for a bug, where pasting to newly created Quill
will not update underlying ngModel.
It also fixes situation when everything is removed from the editor
and it still contains single "<br/>" tag (editor.getLenght()
returns 1). This also fixes errors on Ctrl+A and removing editor
contents.